refactored code to add scripts for CO
