+++ AppSprintPerformanceTests/CICT/UserInputs/ny_cict_user_inputs.py
