+++ AppSprintPerformanceTests/CICT/WorkflowActions/workflows/ny_hq_actions.py
@@ -5,7 +5,7 @@ from selenium.common.exceptions import TimeoutException
 from selenium.webdriver.common.by import By
 
 from common_utilities.selenium.base_page import BasePage
-from PerformanceTests.NY_CICT_App_Sprint.UserInputs.user_inputs import UserData
+from AppSprintPerformanceTests.CICT.UserInputs.ny_cict_user_inputs import UserData
 from common_utilities.decorators import timer
 
 """"Contains test page elements and functions related to workflows in test"""
